devices/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
psycopg2-binary
python-dotenv
boto3
pydantic
pydantic-settings
httpx
